16020000034 Aval. Desempenho > Movimentações > Responder Avaliações de Desempenho: Ajuste no cálculo da performance da competência de acordo com o percentual obtido pelos comportamentos.

+++ src/com/fortes/rh/business/captacao/NivelCompetenciaManager.java
@@ -17,7 +17,7 @@ public interface NivelCompetenciaManager extends GenericManager<NivelCompetencia
 
 	Collection<ConfiguracaoNivelCompetencia> findByCargoOrEmpresa(Long id, Long empresaId);
 
-	Integer getPontuacaoObtidaByConfiguracoesNiveisCompetencia(Collection<ConfiguracaoNivelCompetencia> niveisCompetenciaMarcados);
+	Double getPontuacaoObtidaByConfiguracoesNiveisCompetencia(Collection<ConfiguracaoNivelCompetencia> niveisCompetenciaMarcados);
 
 	int getOrdemMaxima(Long empresaId, Date data);
